+++ src/plugins/fisheye/index.ts
@@ -1,6 +1,6 @@
 import createDOM from '@antv/dom-util/lib/create-dom';
 import modifyCSS from '@antv/dom-util/lib/modify-css';
-import { IG6GraphEvent } from '../../types';
+import { IG6GraphEvent, ShapeStyle } from '../../types';
 import { IGraph } from '../../interface/graph';
 import Graph from '../../graph/graph';
 import Base from '../base';
@@ -8,7 +8,17 @@ import Base from '../base';
 interface FisheyeConfig {
   trigger?: 'mousemove' | 'click',
   d?: number,
-  r?: number
+  r?: number,
+  delegateStyle?: ShapeStyle,
+  showLabel?: boolean
+}
+
+const lensDelegateStyle = {
+  stroke: '#000',
+  strokeOpacity: 0.8,
+  lineWidth: 2,
+  fillOpacity: 0.1,
+  fill: '#ccc'
 }
 export default class Fisheye extends Base {
   constructor(cfg?: FisheyeConfig) {
@@ -18,7 +28,9 @@ export default class Fisheye extends Base {
     return {
       trigger: 'mousemove',
       d: 1.5,
-      r: 100
+      r: 300,
+      delegateStyle: lensDelegateStyle,
+      showLabel: false
     };
   }
 
@@ -53,12 +65,14 @@ export default class Fisheye extends Base {
     const graph: Graph = self.get('graph');
     const cachedMagnifiedModels = self.get('cachedMagnifiedModels');
     const cachedOriginPositions = self.get('cachedOriginPositions');
+    const showLabel = self.get('showLabel');
     const r = self.get('r');
     const r2 = self.get('r2');
     const d = self.get('d');
     const nodes = graph.getNodes();
     const nodeLength = nodes.length;
     const mCenter = { x: e.x, y: e.y };
+    self.updateDelegate(mCenter, r);
     for (let i = 0; i < nodeLength; i++) {
       const model = nodes[i].getModel();
       const { x, y } = model;
@@ -73,9 +87,26 @@ export default class Fisheye extends Base {
         model.x = cos * magnifiedDist + mCenter.x;
         model.y = sin * magnifiedDist + mCenter.y;
         if (!cachedOriginPositions[model.id]) {
-          cachedOriginPositions[model.id] = { x, y };
+          cachedOriginPositions[model.id] = { x, y, texts: [] };
         }
         cachedMagnifiedModels.push(model);
+        if (showLabel && 2 * dist < r) {
+          const node = nodes[i];
+          const nodeGroup = node.getContainer();
+          const shapes = nodeGroup.getChildren();
+          const shapeLength = shapes.length;
+          for (let j = 0; j < shapeLength; j++) {
+            const shape = shapes[j];
+            if (shape.get('type') === 'text') {
+              cachedOriginPositions[model.id].texts.push({
+                visible: shape.get('visible'),
+                shape
+              });
+              shape.set('visible', true);
+            }
+          }
+
+        }
       }
     }
     graph.refreshPositions();
@@ -92,8 +123,12 @@ export default class Fisheye extends Base {
     for (let i = 0; i < cacheLength; i++) {
       const node = cachedMagnifiedModels[i];
       const id = node.id;
-      node.x = cachedOriginPositions[id].x;
-      node.y = cachedOriginPositions[id].y;
+      const ori = cachedOriginPositions[id];
+      node.x = ori.x;
+      node.y = ori.y;
+      ori.texts.forEach(text => {
+        text.shape.set('visible', text.visible);
+      });
     }
     self.set('cachedMagnifiedModels', []);
     self.set('cachedOriginPositions', {});
@@ -114,10 +149,47 @@ export default class Fisheye extends Base {
     }
   }
 
+
+  /**
+   * 放大镜的图形
+   * @param {Point} mCenter
+   * @param {number} r
+   */
+  private updateDelegate(mCenter, r) {
+    const self = this;
+    const graph = self.get('graph');
+    let lensDelegate = self.get('delegate');
+    if (!lensDelegate || lensDelegate.destroyed) {
+      // 拖动多个
+      const parent = graph.get('group');
+      const attrs = self.get('delegateStyle') || lensDelegateStyle;
+
+      // model上的x, y是相对于图形中心的，delegateShape是g实例，x,y是绝对坐标
+      lensDelegate = parent.addShape('circle', {
+        attrs: {
+          r: r / 1.5,
+          x: mCenter.x,
+          y: mCenter.y,
+          ...attrs,
+        },
+        name: 'lens-delegate-shape',
+      });
+      lensDelegate.set('capture', false);
+    } else {
+      lensDelegate.attr({
+        x: mCenter.x,
+        y: mCenter.y,
+      });
+    }
+    self.set('delegate', lensDelegate);
+  }
+
   public clear() {
     const graph = this.get('graph');
     this.restoreCache();
     graph.refreshPositions();
+    let lensDelegate = this.get('delegate');
+    lensDelegate.destroy();
 
   }
 }