+++ stories/Interaction/component/fisheye-lens.tsx
@@ -7,20 +7,34 @@ let graph: IGraph = null;
 const FishEye = () => {
   const container = React.useRef();
   const fisheye = new G6.Fisheye({
-    r: 100
+    r: 200,
+    showLabel: true,
   });
+  const colors = [
+    '#8FE9FF',
+    '#87EAEF',
+    '#FFC9E3',
+    '#A7C2FF',
+    '#FFA1E3',
+    '#FFE269',
+    '#BFCFEE',
+    '#FFA0C5',
+    '#D5FF86',
+  ]
+
   useEffect(() => {
     if (!graph) {
       graph = new G6.Graph({
         container: container.current as string | HTMLElement,
-        width: 500,
-        height: 500,
+        width: 1000,
+        height: 700,
         plugins: [fisheye],
         layout: {
-          type: 'force'
+          type: 'force',
+          preventOverlap: true
         },
         modes: {
-          default: ['drag-canvas']
+          default: ['drag-canvas', 'zoom-canvas']
         }
       });
       fetch('https://gw.alipayobjects.com/os/antvdemo/assets/data/relations.json')
@@ -28,9 +42,20 @@ const FishEye = () => {
         .then(data => {
           data.nodes.forEach(node => {
             node.label = node.id;
+            node.size = Math.random() * 40 + 20;
+            node.style = {
+              fill: colors[Math.floor(Math.random() * 9)],
+              lineWidth: 0
+            }
+            console.log(node)
           });
           graph.data(data);
           graph.render();
+          graph.getNodes().forEach(node => {
+            node.getContainer().getChildren().forEach(shape => {
+              if (shape.get('type') === 'text') shape.set('visible', false);
+            });
+          });
         });
     }
   });