+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # ChangeLog
 
+#### 3.5
+- feat: fisheye lens plugin.
+
 #### 3.5.12
 - fix: node:click is triggered twice while clicking a node;
 - fix: update combo edge when drag node out of it problem;